New release (1.2) without the class scanner, should work everywhere now
Also I totally meant this to be an extra commit and did not forget to
include this in the previous one.

Yes I do know about -amend, but I have nothing better to do than writing
this.

This is probably the longest commit message you will ever see from me.